functions_python/requirements.txt
firebase-functions
firebase-admin
pandas
numpy
requests
PyPortfolioOpt
cvxopt
scikit-learn
scipy
google-generativeai